requirements.txt
python-telegram-bot
logging
requests

